actually works now

+++ lib/capistrano/transmit.rb
@@ -1,47 +1,53 @@
-after "transmit:get:mysql", "transmit:cleanup"
-after "transmit:put:mysql", "transmit:cleanup"
+unless Capistrano::Configuration.respond_to?(:instance)
+  abort "capistrano/transmit requires Capistrano 2"
+end
 
-namespace :transmit do
-  namespace :get do
-    desc 'Fetch the remote production database and overwrite your local development database with it'
-    task :mysql, :roles => :db do
-      run "mysqldump --opt --quick --extended-insert --skip-lock-tables -u #{db_remote['username']} --password='#{db_remote['password']}' --host='#{db_remote['host']}' #{db_remote['database']} | gzip > #{dumpfile}"
+Capistrano::Configuration.instance.load do
+  after "transmit:get:mysql", "transmit:cleanup"
+  after "transmit:put:mysql", "transmit:cleanup"
 
-      system "rsync -vP #{user}@#{deploy_host}:#{dumpfile} tmp/#{db_local["database"]}.sql.gz"
-      system "gunzip < tmp/#{db_local["database"]}.sql.gz | mysql -u #{db_local['username']} --password='#{db_local['password']}' --host='#{db_local['host']}' #{db_local['database']}"
+  namespace :transmit do
+    namespace :get do
+      desc 'Fetch the remote production database and overwrite your local development database with it'
+      task :mysql, :roles => :db do
+        run "mysqldump --opt --quick --extended-insert --skip-lock-tables -u #{db_remote['username']} --password='#{db_remote['password']}' --host='#{db_remote['host']}' #{db_remote['database']} | gzip > #{dumpfile}"
+
+        system "rsync -vP #{user}@#{deploy_host}:#{dumpfile} tmp/#{db_local["database"]}.sql.gz"
+        system "gunzip < tmp/#{db_local["database"]}.sql.gz | mysql -u #{db_local['username']} --password='#{db_local['password']}' --host='#{db_local['host']}' #{db_local['database']}"
+      end
+
+      desc 'Fetch the assets from the production server to the development environment'
+      task :assets, :roles => :app do
+        system "rsync -Lcrvz #{user}@#{deploy_host}:#{current_path}/public ."
+      end
     end
 
-    desc 'Fetch the assets from the production server to the development environment'
-    task :assets, :roles => :app do
-      system "rsync -Lcrvz #{user}@#{deploy_host}:#{current_path}/public ."
-    end
-  end
+    namespace :put do
+      desc 'Upload the local development database to the remote production database and overwrite it'
+      task :mysql, :roles => :db do
+        system "mysqldump --opt -u #{db_local['username']} --password='#{db_local['password']}' --host='#{db_local['host']}' #{db_local['database']} > tmp/#{db_local['database']}.sql"
 
-  namespace :put do
-    desc 'Upload the local development database to the remote production database and overwrite it'
-    task :mysql, :roles => :db do
-      system "mysqldump --opt -u #{db_local['username']} --password='#{db_local['password']}' --host='#{db_local['host']}' #{db_local['database']} > tmp/#{db_local['database']}.sql"
-
-      system "rsync -vP tmp/#{db_local['database']}.sql #{user}@#{deploy_host}:#{dumpfile}"
-      run "mysql -u #{db_remote['username']} --password='#{db_remote['password']}' --host='#{db_remote['host']}' #{db_remote['database']} < #{dumpfile}"
+        system "rsync -vP tmp/#{db_local['database']}.sql #{user}@#{deploy_host}:#{dumpfile}"
+        run "mysql -u #{db_remote['username']} --password='#{db_remote['password']}' --host='#{db_remote['host']}' #{db_remote['database']} < #{dumpfile}"
+      end
     end
-  end
   
-  task :cleanup do
-    run "rm #{dumpfile}"
-    system "rm tmp/#{db_local['database']}.sql.gz"
+    task :cleanup do
+      run "rm #{dumpfile}"
+      system "rm tmp/#{db_local['database']}.sql.gz"
+    end
   end
-end
 
-set(:db_remote) do
-  db_config = capture "cat #{current_path}/config/database.yml"
-  YAML::load(db_config)['production']
-end
+  set(:db_remote) do
+    db_config = capture "cat #{current_path}/config/database.yml"
+    YAML::load(db_config)['production']
+  end
 
-set(:db_local) do
-  YAML::load_file("config/database.yml")['development']
-end
+  set(:db_local) do
+    YAML::load_file("config/database.yml")['development']
+  end
 
-set :dumpfile do
-  "#{current_path}/tmp/#{db_remote['database']}.sql.gz"
+  set :dumpfile do
+    "#{current_path}/tmp/#{db_remote['database']}.sql.gz"
+  end
 end
